% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chain.R
\name{chain}
\alias{chain}
\title{Chained execution of a list of functions}
\usage{
chain(l, arg, ...)
}
\arguments{
\item{l}{A list of function to be executed in chain}

\item{arg}{the argument passed to the first function}

\item{...}{further arguments passed to all functions}
}
\value{
The result of the last function in the list
}
\description{
The first function in list \code{l} is executed with the given \code{arg} then the result
is passed to the second function and so on for each function in the list.
}
\author{
Michal Burda
}
