% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauss_rulebase.R
\name{gauss_rulebase}
\alias{gauss_rulebase}
\title{Create the empty Gauss rulebase}
\usage{
gauss_rulebase(n_inputs, n_parts)
}
\arguments{
\item{n_inputs}{the number of input variables}

\item{n_parts}{the number of Gauss fuzzy sets to partition the space of each variable to}
}
\value{
An instance of the \code{gaus_rulebase} S3 class, which is a list with the following elements:
- n_inputs - the number of input variables
- n_parts - the number of fuzzy sets in the partitioning of the space of each input variable
- means - an $(n_inputs, n_parts)$ matrix of the $m$ parameters of the Gauss distribution, for
input $i$ and part $j$
- stdevs - an $(n_inputs, n_parts)$ matrix of the $\sigma$ parameters of the Gauss distribution,
for input $i$ and part $j$
- antecedents - an $(k, n_inputs)$ matrix of combinations of parts assigned to inputs
- consequents - an $(n_inputs + 1, k)$ matrix of linear function\preformatted{Here $k = n_parts^\{n_inputs\}$ stands for the number of rules in the rulebase.
}
}
\description{
Create the empty Gauss rulebase
}
\author{
Michal Burda
}
