% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ig_split.R
\name{ig_split}
\alias{ig_split}
\title{Thee node split based on the Information Gain (IG)}
\usage{
ig_split(data, cfg)
}
\value{

}
\description{
Thee node split based on the Information Gain (IG)
}
\seealso{
https://scientistcafe.com/ids/splitting-criteria.html
}
\author{
Michal Burda
}
