% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igr.R
\name{igr}
\alias{igr}
\title{Compute the Information Gain Ratio (IGR)}
\usage{
igr(x1, x2)
}
\description{
Compute the Information Gain Ratio (IGR)
}
\seealso{
https://scientistcafe.com/ids/splitting-criteria.html
}
\author{
Michal Burda
}
