% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igr.R
\name{igr_by_indices}
\alias{igr_by_indices}
\title{Compute the Information Gain Ratio (IGR)}
\usage{
igr_by_indices(x, indices)
}
\value{

}
\description{
Compute the Information Gain Ratio (IGR)
}
\seealso{
https://scientistcafe.com/ids/splitting-criteria.html
}
\author{
Michal Burda
}
