% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igr_split.R
\name{igr_split}
\alias{igr_split}
\title{Thee node split based on the Information Gain Ratio (IGR)}
\usage{
igr_split(data, cfg)
}
\value{

}
\description{
Thee node split based on the Information Gain Ratio (IGR)
}
\seealso{
https://scientistcafe.com/ids/splitting-criteria.html
}
\author{
Michal Burda
}
