% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_result.R
\name{mean_result}
\alias{mean_result}
\title{Create the leaf node result by obtaining the mean value}
\usage{
mean_result(data)
}
\arguments{
\item{data}{the instance of the \code{woods_data} S3 class}
}
\value{
the list with the mean as \code{value} and \code{"mean"} as \code{label}
}
\description{
Create the leaf node result by obtaining the mean value
}
\author{
Michal Burda
}
