% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode_result.R
\name{mode_result}
\alias{mode_result}
\title{Create the leaf node result by obtaining the mode, i.e., the most frequent
value in data}
\usage{
mode_result(data)
}
\arguments{
\item{data}{the instance of the \code{woods_data} S3 class}
}
\value{
the list with the mode as \code{value} and \code{"mode"} as \code{label}
}
\description{
Create the leaf node result by obtaining the mode, i.e., the most frequent
value in data
}
\author{
Michal Burda
}
