% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modus.R
\name{modus}
\alias{modus}
\title{Get the most frequent value}
\usage{
modus(x, na.rm = FALSE)
}
\value{
The modus (mode), i.e., the most frequent value
}
\description{
Get the most frequent value
}
\author{
Michal Burda
}
