% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-cutpoint_condition.R
\name{predict.cutpoint_condition}
\alias{predict.cutpoint_condition}
\title{Returns vector of logical values indicating satisfiability of the condition in data rows}
\usage{
\method{predict}{cutpoint_condition}(object, newdata, ...)
}
\description{
Returns vector of logical values indicating satisfiability of the condition in data rows
}
\author{
Michal Burda
}
