% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling_factory.R
\name{resampling_factory}
\alias{resampling_factory}
\title{Create a function for row-wise and col-wise resampling of data}
\usage{
resampling_factory(rows = NULL, cols = NULL)
}
\arguments{
\item{rows}{either a NULL value indicating no resampling should be done at all,
or a positive value indicating the required number of rows
or a function that from a number of rows of input data computes the required number of
rows to be selected (i.e., identity function passed here causes the resampling with
replacement of all data rows).}

\item{cols}{either a NULL value indicating no resampling should be done at all,
or a positive value indicating the required number of columns
or a function that from a number of columns of input data computes the required number of
columns to be selected}
}
\value{
a single-argument function that takes an instance of the \code{woods_data} S3 class and
resamples its values to satisfy the number of rows and columns as specified by the \code{rows}
and \code{cols} arguments described above.
}
\description{
For rows and columns, the resampling is performed with and without replacement, respectively.
}
\author{
Michal Burda
}
