% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split-default.R
\name{split.default}
\alias{split.default}
\title{Split data to \code{left} and \code{right} part accordingly to the logical vector}
\usage{
\method{split}{default}(def, data, cfg)
}
\arguments{
\item{def}{a logical vector indicating which data rows belong to the \code{left} part of the split}

\item{data}{the instance of the \code{woods_data} S3 class}

\item{cfg}{the configuration list (not needed for this function)}
}
\value{
a list with \code{left} and \code{right} data (\code{left} data are those indicated as \code{TRUE} in the
\code{def} vector, \code{right} data correspond to \code{FALSE} in the \code{def} vector)
}
\description{
Split data to \code{left} and \code{right} part accordingly to the logical vector
}
\author{
Michal Burda
}
