% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sse.R
\name{sse}
\alias{sse}
\title{Compute Sum of Squared Error (SSE)}
\usage{
sse(x)
}
\value{

}
\description{
Compute Sum of Squared Error (SSE)
}
\seealso{
https://scientistcafe.com/ids/splitting-criteria.html
}
\author{
Michal Burda
}
