% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sse_split.R
\name{sse_split}
\alias{sse_split}
\title{Tree node split based on the Sum of Squared Error (SSE)}
\usage{
sse_split(data, cfg)
}
\value{

}
\description{
Tree node split based on the Sum of Squared Error (SSE)
}
\seealso{
https://scientistcafe.com/ids/splitting-criteria.html
}
\author{
Michal Burda
}
