% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woods.R
\name{woods.default}
\alias{woods.default}
\title{Train the double random forest}
\usage{
\method{woods}{default}(
  y,
  x,
  n_tree = 2,
  mtry = ceiling(sqrt(ncol(x))),
  max_height = NA,
  node_size = 1
)
}
\value{
the double random forest model
}
\description{
Train the double random forest
}
\author{
Michal Burda
}
