% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woods-formula.R
\name{woods.formula}
\alias{woods.formula}
\title{Train the double random forest}
\usage{
\method{woods}{formula}(formula, data, subset = NULL, na.action = na.fail, ...)
}
\value{
the double random forest model
}
\description{
Train the double random forest
}
\author{
Michal Burda
}
