% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woods_data.R
\name{woods_data}
\alias{woods_data}
\title{Create the data structure used internally in the \code{\link[=woods]{woods()}} algorithm.}
\usage{
woods_data(y, x)
}
\arguments{
\item{y}{the response vector}

\item{x}{the independent variables as data.frame}
}
\value{
the S3 object of class \code{woods_data}
}
\description{
Create the data structure used internally in the \code{\link[=woods]{woods()}} algorithm.
}
\author{
Michal Burda
}
