% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/woods.R
\name{woods.default}
\alias{woods.default}
\title{Train the double random forest}
\usage{
\method{woods}{default}(
  y,
  x,
  n_tree = 500,
  mtry = if (is.factor(y)) floor(sqrt(ncol(x))) else max(floor(ncol(x)/3), 1),
  max_height = NA,
  node_size = if (is.factor(y)) 1 else 5,
  resample_rows = TRUE,
  principal_component = TRUE,
  linear_model = TRUE,
  hypersphere = TRUE
)
}
\value{
the double random forest model
}
\description{
Train the double random forest
}
\author{
Michal Burda
}
